matc/requirements.txt
Django~=4.0.5
django-cors-headers~=3.13.0
djangorestframework~=3.13.1
djangorestframework-simplejwt~=5.2.0
drf-yasg~=1.20.0
Pillow~=9.1.1
urllib3==1.25.7